requirements.txt
Django==1.10.2
djangorestframework==3.5.1
pygame==1.9.2b6
requests==2.11.1
